04_TrabajosWeb_Scraper/requirements.txt
selenium
psycopg2-binary
python-dotenv
requests